Jwt login and authentication system entegration

+++ goldendaysservice/src/main/java/com/goldendays/contract/IDummyService.java
@@ -1,5 +1,6 @@
 package com.goldendays.contract;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.goldendays.contract.modelobjects.Person;
@@ -8,11 +9,11 @@ public interface IDummyService {
 
 	public void addPerson(Person person);
 
-	public void editPerson(Person person, int personId);
+	public void editPerson(Person person, BigDecimal personId);
 
-	public void deletePerson(int personId);
+	public void deletePerson(BigDecimal personId);
 
-	public Person find(int personId);
+	public Person find(BigDecimal personId);
 
 	public List<Person> findAll();
 }
